src/api/requirements.txt
aniso8601==7.0.0
audioread==2.1.8
Babel==2.8.0
bcrypt==3.1.7
blinker==1.4
cffi==1.13.2
Click==7.0
decorator==4.4.1
Django==2.2.7
django-filter==2.2.0
Flask==1.1.1
Flask-BabelEx==0.9.3
Flask-Bcrypt==0.7.1
Flask-Cors==3.0.8
Flask-GraphQL==2.0.0
Flask-GraphQL-Auth==1.3.0
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Principal==0.4.0
Flask-Security==3.0.0
Flask-WTF==0.14.2
graphene==2.1.8
graphene-file-upload==1.2.2
graphene-sqlalchemy==2.2.2
graphene-sqlalchemy-filter==1.10.0
graphql-core==2.2.1
graphql-relay==2.0.0
graphql-server-core==1.1.1
greenlet==0.4.15
iso8601==0.1.12
itsdangerous==1.1.0
Jinja2==2.10.3
joblib==0.14.1
librosa==0.7.1
llvmlite==0.30.0
MarkupSafe==1.1.1
numba==0.46.0
numpy==1.17.4
passlib==1.7.2
pendulum==2.0.5
promise==2.2.1
pycparser==2.19
PyJWT==1.7.1
PyMySQL==0.9.3
python-dateutil==2.8.1
pytz==2019.3
pytzdata==2019.3
resampy==0.2.2
Rx==1.6.1
scikit-learn==0.22
scipy==1.3.3
singledispatch==3.4.0.3
six==1.13.0
SoundFile==0.10.3.post1
speaklater==1.3
SQLAlchemy==1.3.11
sqlparse==0.3.0
Werkzeug==0.16.0
WTForms==2.2.1
